+++ datafusion/sql/src/lib.rs
@@ -41,6 +41,7 @@ pub mod parser;
 pub mod planner;
 mod query;
 mod relation;
+pub mod resolve;
 mod select;
 mod set_expr;
 mod stack;
@@ -49,6 +50,9 @@ mod statement;
 pub mod unparser;
 pub mod utils;
 mod values;
-
+#[deprecated(
+    since = "46.0.0",
+    note = "use datafusion_common::{ResolvedTableReference, TableReference}"
+)]
 pub use datafusion_common::{ResolvedTableReference, TableReference};
 pub use sqlparser;
